Validation Dsl created #73
