+++ buildSrc/src/main/kotlin/BuildConstants.kt
@@ -1,0 +1,9 @@
+object TestFolders {
+    const val Integration = "test-int"
+    const val e2e = "test-e2e"
+}
+
+object TaskNames {
+    const val testAggregateReports = "testAggregateReports"
+    const val codeCoverageReport = "codeCoverageReport"
+}
