--- buildSrc/src/main/kotlin/coverage.gradle.kts
@@ -1,19 +1,0 @@
-plugins {
-    base
-    `jacoco-report-aggregation`
-    `test-report-aggregation`
-}
-
-dependencies {
-    jacocoAggregation(project(":lib:stove-testing-e2e"))
-    jacocoAggregation(project(":lib:stove-testing-e2e-couchbase"))
-    jacocoAggregation(project(":lib:stove-testing-e2e-elasticsearch"))
-    jacocoAggregation(project(":lib:stove-testing-e2e-http"))
-    jacocoAggregation(project(":lib:stove-testing-e2e-kafka"))
-    jacocoAggregation(project(":lib:stove-testing-e2e-wiremock"))
-    jacocoAggregation(project(":lib:stove-testing-e2e-rdbms"))
-    jacocoAggregation(project(":lib:stove-testing-e2e-rdbms-postgres"))
-    jacocoAggregation(project(":starters:spring:stove-spring-testing-e2e"))
-    jacocoAggregation(project(":starters:spring:stove-spring-testing-e2e-kafka"))
-    jacocoAggregation(project(":starters:ktor:stove-ktor-testing-e2e"))
-}
