+++ lib/stove-testing-e2e-couchbase/src/main/kotlin/com/trendyol/stove/testing/e2e/couchbase/Extensions.kt
@@ -3,6 +3,7 @@ package com.trendyol.stove.testing.e2e.couchbase
 import arrow.core.getOrElse
 import com.trendyol.stove.testing.e2e.containers.withProvidedRegistry
 import com.trendyol.stove.testing.e2e.system.TestSystem
+import com.trendyol.stove.testing.e2e.system.ValidationDsl
 import com.trendyol.stove.testing.e2e.system.abstractions.SystemNotRegisteredException
 import org.testcontainers.couchbase.BucketDefinition
 import org.testcontainers.couchbase.CouchbaseContainer
@@ -26,3 +27,6 @@ fun TestSystem.couchbase(): CouchbaseSystem =
     getOrNone<CouchbaseSystem>().getOrElse {
         throw SystemNotRegisteredException(CouchbaseSystem::class)
     }
+
+suspend fun ValidationDsl.couchbase(validation: suspend CouchbaseSystem.() -> Unit): Unit =
+    validation(this.testSystem.couchbase())
