+++ lib/stove-testing-e2e-couchbase/src/test/kotlin/com/trendyol/stove/testing/e2e/couchbase/CouchbaseTestSystemTests.kt
@@ -6,6 +6,7 @@ import com.trendyol.stove.testing.e2e.couchbase.CouchbaseSystem.Companion.should
 import com.trendyol.stove.testing.e2e.database.DocumentDatabaseSystem.Companion.shouldGet
 import com.trendyol.stove.testing.e2e.database.migrations.DatabaseMigration
 import com.trendyol.stove.testing.e2e.system.TestSystem
+import com.trendyol.stove.testing.e2e.system.TestSystem.Companion.validate
 import com.trendyol.stove.testing.e2e.system.abstractions.ApplicationUnderTest
 import io.kotest.core.config.AbstractProjectConfig
 import io.kotest.core.spec.style.FunSpec
@@ -14,7 +15,7 @@ import kotlinx.coroutines.reactive.awaitFirstOrNull
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import java.time.Duration
-import java.util.UUID
+import java.util.*
 
 const val testBucket = "test-couchbase-bucket"
 
@@ -77,5 +78,32 @@ class CouchbaseTestSystemTests : FunSpec({
                 actual.id shouldBe id
                 actual.description shouldBe testCase.name.testName
             }
+    }
+})
+
+class CouchbaseTestSystemUsesDslTests : FunSpec({
+
+    data class ExampleInstance(
+        val id: String,
+        val description: String,
+    )
+
+    test("should save and get") {
+        val id = UUID.randomUUID().toString()
+        val anotherCollectionName = "another"
+        validate {
+            couchbase {
+                saveToDefaultCollection(id, ExampleInstance(id = id, description = testCase.name.testName))
+                save(anotherCollectionName, id = id, ExampleInstance(id = id, description = testCase.name.testName))
+                shouldGet<ExampleInstance>(id) { actual ->
+                    actual.id shouldBe id
+                    actual.description shouldBe testCase.name.testName
+                }
+                shouldGet<ExampleInstance>(anotherCollectionName, id) { actual ->
+                    actual.id shouldBe id
+                    actual.description shouldBe testCase.name.testName
+                }
+            }
+        }
     }
 })
