+++ lib/stove-testing-e2e-http/src/main/kotlin/com/trendyol/stove/testing/e2e/http/DefaultHttpSystem.kt
@@ -8,6 +8,7 @@ import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.kotlin.readValue
 import com.trendyol.stove.testing.e2e.serialization.StoveObjectMapper
 import com.trendyol.stove.testing.e2e.system.TestSystem
+import com.trendyol.stove.testing.e2e.system.ValidationDsl
 import com.trendyol.stove.testing.e2e.system.abstractions.SystemNotRegisteredException
 import com.trendyol.stove.testing.e2e.system.abstractions.SystemOptions
 import kotlinx.coroutines.future.await
@@ -52,6 +53,9 @@ fun TestSystem.http(): DefaultHttpSystem =
     getOrNone<DefaultHttpSystem>().getOrElse {
         throw SystemNotRegisteredException(DefaultHttpSystem::class)
     }
+
+suspend fun ValidationDsl.http(validation: suspend DefaultHttpSystem.() -> Unit): Unit =
+    validation(this.testSystem.http())
 
 class DefaultHttpSystem(
     override val testSystem: TestSystem,
