+++ lib/stove-testing-e2e-mongodb/src/main/kotlin/com/trendyol/stove/testing/e2e/mongodb/Options.kt
@@ -1,8 +1,13 @@
 package com.trendyol.stove.testing.e2e.mongodb
 
+import arrow.core.getOrElse
 import com.trendyol.stove.testing.e2e.containers.ContainerOptions
 import com.trendyol.stove.testing.e2e.containers.DEFAULT_REGISTRY
+import com.trendyol.stove.testing.e2e.containers.withProvidedRegistry
+import com.trendyol.stove.testing.e2e.system.TestSystem
+import com.trendyol.stove.testing.e2e.system.ValidationDsl
 import com.trendyol.stove.testing.e2e.system.abstractions.ExposedConfiguration
+import com.trendyol.stove.testing.e2e.system.abstractions.SystemNotRegisteredException
 import org.testcontainers.containers.MongoDBContainer
 
 data class MongodbExposedConfiguration(
@@ -30,3 +35,29 @@ data class DatabaseOptions(
         val collection: String = "stoveCollection",
     )
 }
+
+fun TestSystem.withMongodb(
+    options: MongodbSystemOptions,
+): TestSystem {
+    val mongodbContainer =
+        withProvidedRegistry(options.container.imageWithTag, options.container.registry) {
+            MongoDBContainer(it)
+                .withReuse(this.options.keepDependenciesRunning)
+        }
+    this.getOrRegister(
+        MongodbSystem(this, MongodbContext(mongodbContainer, options))
+    )
+    return this
+}
+
+fun TestSystem.withMongodb(
+    optionsFn: () -> MongodbSystemOptions,
+): TestSystem = withMongodb(optionsFn())
+
+fun TestSystem.mongodb(): MongodbSystem =
+    getOrNone<MongodbSystem>().getOrElse {
+        throw SystemNotRegisteredException(MongodbSystem::class)
+    }
+
+suspend fun ValidationDsl.mongodb(validation: suspend MongodbSystem.() -> Unit): Unit =
+    validation(this.testSystem.mongodb())
