+++ lib/stove-testing-e2e/src/main/kotlin/com/trendyol/stove/testing/e2e/system/TestSystem.kt
@@ -76,6 +76,8 @@ class TestSystem(
          * DO NOT access it before [run] completes
          */
         lateinit var instance: TestSystem
+
+        suspend fun validate(validation: suspend ValidationDsl.() -> Unit): Unit = validation(ValidationDsl(instance))
     }
 
     /**
