+++ lib/stove-testing-e2e/src/main/kotlin/com/trendyol/stove/testing/e2e/system/ValidationDsl.kt
@@ -1,0 +1,4 @@
+package com.trendyol.stove.testing.e2e.system
+
+@JvmInline
+value class ValidationDsl(val testSystem: TestSystem)
