+++ starters/spring/stove-spring-testing-e2e-kafka/src/main/kotlin/com/trendyol/stove/testing/e2e/kafka/KafkaSystem.kt
@@ -3,16 +3,14 @@ package com.trendyol.stove.testing.e2e.kafka
 import arrow.core.Option
 import arrow.core.Some
 import arrow.core.getOrElse
-import com.fasterxml.jackson.databind.ObjectMapper
 import com.trendyol.stove.functional.Try
 import com.trendyol.stove.functional.recover
-import com.trendyol.stove.testing.e2e.containers.DEFAULT_REGISTRY
-import com.trendyol.stove.testing.e2e.containers.withProvidedRegistry
 import com.trendyol.stove.testing.e2e.messaging.AssertsPublishing
 import com.trendyol.stove.testing.e2e.messaging.MessagingSystem
-import com.trendyol.stove.testing.e2e.serialization.StoveObjectMapper
 import com.trendyol.stove.testing.e2e.system.TestSystem
-import com.trendyol.stove.testing.e2e.system.abstractions.*
+import com.trendyol.stove.testing.e2e.system.abstractions.ExposesConfiguration
+import com.trendyol.stove.testing.e2e.system.abstractions.RunnableSystemWithContext
+import com.trendyol.stove.testing.e2e.system.abstractions.StateOfSystem
 import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.future.await
 import kotlinx.coroutines.runBlocking
@@ -23,38 +21,8 @@ import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.getBean
 import org.springframework.context.ApplicationContext
 import org.springframework.kafka.core.KafkaTemplate
-import org.testcontainers.containers.KafkaContainer
 import kotlin.reflect.KClass
 import kotlin.time.Duration
-
-data class KafkaExposedConfiguration(
-    val bootstrapServers: String,
-) : ExposedConfiguration
-
-data class KafkaSystemOptions(
-    val registry: String = DEFAULT_REGISTRY,
-    val ports: List<Int> = listOf(9092, 9093),
-    val objectMapper: ObjectMapper = StoveObjectMapper.Default,
-    override val configureExposedConfiguration: (KafkaExposedConfiguration) -> List<String> = { _ -> listOf() },
-) : SystemOptions, ConfiguresExposedConfiguration<KafkaExposedConfiguration>
-
-data class KafkaContext(
-    val container: KafkaContainer,
-    val objectMapper: ObjectMapper,
-    val configureExposedConfiguration: (KafkaExposedConfiguration) -> List<String>,
-)
-
-fun TestSystem.withKafka(
-    options: KafkaSystemOptions = KafkaSystemOptions(),
-): TestSystem = withProvidedRegistry("confluentinc/cp-kafka:latest", options.registry) {
-    KafkaContainer(it).withExposedPorts(*options.ports.toTypedArray())
-        .withEmbeddedZookeeper()
-        .withReuse(this.options.keepDependenciesRunning)
-}.let { getOrRegister(KafkaSystem(this, KafkaContext(it, options.objectMapper, options.configureExposedConfiguration))) }
-    .let { this }
-
-fun TestSystem.kafka(): KafkaSystem =
-    getOrNone<KafkaSystem>().getOrElse { throw SystemNotRegisteredException(KafkaSystem::class) }
 
 class KafkaSystem(
     override val testSystem: TestSystem,
